# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Snyder Dialer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface UserRegister {
  email: string;
  password: string;
  name: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface DashboardStats {
  totalCampaigns: number;
  activeCampaigns: number;
  totalLeads: number;
  totalCalls: number;
  successfulCalls: number;
  avgCallDuration: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  call_completed: 'call_completed',
  campaign_launched: 'campaign_launched',
  leads_uploaded: 'leads_uploaded',
  campaign_paused: 'campaign_paused',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  /** @nullable */
  campaignId?: number | null;
  /** @nullable */
  campaignName?: string | null;
  createdAt: string;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface Campaign {
  id: number;
  name: string;
  status: CampaignStatus;
  objective: string;
  masterPrompt: string;
  /** @nullable */
  twilioAccountSid?: string | null;
  /** @nullable */
  twilioAuthToken?: string | null;
  /** @nullable */
  twilioPhoneNumber?: string | null;
  /** @nullable */
  vapiApiKey?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CampaignInput {
  /** @minLength 1 */
  name: string;
  objective: string;
  masterPrompt: string;
  twilioAccountSid?: string;
  twilioAuthToken?: string;
  twilioPhoneNumber?: string;
  vapiApiKey?: string;
}

export interface CampaignUpdate {
  /** @minLength 1 */
  name?: string;
  objective?: string;
  masterPrompt?: string;
  twilioAccountSid?: string;
  twilioAuthToken?: string;
  twilioPhoneNumber?: string;
  vapiApiKey?: string;
}

export interface CampaignStats {
  campaignId: number;
  totalLeads: number;
  calledLeads: number;
  pendingLeads: number;
  successfulCalls: number;
  failedCalls: number;
  avgCallDuration: number;
}

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  pending: 'pending',
  calling: 'calling',
  completed: 'completed',
  failed: 'failed',
  do_not_call: 'do_not_call',
} as const;

export interface Lead {
  id: number;
  campaignId: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  notes?: string | null;
  status: LeadStatus;
  createdAt: string;
}

export interface LeadsCsvUpload {
  file?: Blob;
}

export interface UploadResult {
  imported: number;
  skipped: number;
  total: number;
}

export type KnowledgeBaseFileStatus = typeof KnowledgeBaseFileStatus[keyof typeof KnowledgeBaseFileStatus];


export const KnowledgeBaseFileStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface KnowledgeBaseFile {
  id: number;
  campaignId: number;
  fileName: string;
  fileType: string;
  status: KnowledgeBaseFileStatus;
  createdAt: string;
}

export interface KbDocUpload {
  file?: Blob;
}

export type CallLogStatus = typeof CallLogStatus[keyof typeof CallLogStatus];


export const CallLogStatus = {
  initiated: 'initiated',
  in_progress: 'in_progress',
  completed: 'completed',
  failed: 'failed',
  no_answer: 'no_answer',
  voicemail: 'voicemail',
} as const;

export interface CallLog {
  id: number;
  campaignId: number;
  leadId: number;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  leadPhone?: string | null;
  status: CallLogStatus;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  createdAt: string;
}

export interface PromptTemplate {
  id: number;
  name: string;
  content: string;
  variables?: string[];
  createdAt: string;
}

export interface PromptTemplateInput {
  /** @minLength 1 */
  name: string;
  content: string;
  variables?: string[];
}

export interface PromptTemplateUpdate {
  /** @minLength 1 */
  name?: string;
  content?: string;
  variables?: string[];
}