# lib/api-zod/src/generated/types/callLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Snyder Dialer API
 * OpenAPI spec version: 0.1.0
 */
import type { CallLogStatus } from './callLogStatus';

export interface CallLog {
  id: number;
  campaignId: number;
  leadId: number;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  leadPhone?: string | null;
  status: CallLogStatus;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  startedAt?: Date | null;
  /** @nullable */
  endedAt?: Date | null;
  createdAt: Date;
}