# lib/api-zod/src/generated/types/knowledgeBaseFile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Snyder Dialer API
 * OpenAPI spec version: 0.1.0
 */
import type { KnowledgeBaseFileStatus } from './knowledgeBaseFileStatus';

export interface KnowledgeBaseFile {
  id: number;
  campaignId: number;
  fileName: string;
  fileType: string;
  status: KnowledgeBaseFileStatus;
  createdAt: Date;
}