# lib/api-zod/src/generated/types/userRegister.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Snyder Dialer API
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegister {
  email: string;
  password: string;
  name: string;
}